% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{combine_ests}
\alias{combine_ests}
\title{Combine output from the contigs pipeline and samples pipeline}
\usage{
combine_ests(contigs, samples)
}
\arguments{
\item{contigs}{output from the contigs pipeline}

\item{samples}{output from the samples pipeline}
}
\value{
a dataframe with the combined estimated PTRs
}
\description{
Combine output from the contigs pipeline and samples pipeline
}
