% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{compare_contig_subsets}
\alias{compare_contig_subsets}
\title{Compares contig subset x against contig subset y}
\usage{
compare_contig_subsets(
  est_ptrs_x,
  est_ptrs_y,
  pipeline_x,
  pipeline_y,
  cor_cutoff,
  max_cor
)
}
\arguments{
\item{est_ptrs_x}{PTR estimates from contig subset x}

\item{est_ptrs_y}{PTR estimates from contig subset y}

\item{pipeline_x}{pipeline for contig subset x}

\item{pipeline_y}{pipeline for contig subset y}

\item{cor_cutoff}{the correlation cutoff}

\item{max_cor}{the max correlation}
}
\value{
a named list including the est_ptr dataframe and a max_cor value
\itemize{
\item sample: sample
\item est_ptr: PTR estimate
\item coefficient: coefficient of linear regression
\item pValue: p-value of linear regression
\item cor: correlation coefficient
\item correctY: corrected coverage
}
max_cor: the max correlation achieved
}
\description{
Compares contig subset x against contig subset y
}
