% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{compare_sample_subsets}
\alias{compare_sample_subsets}
\title{Compares sample subset x against sample subset y}
\usage{
compare_sample_subsets(
  est_ptrs_x,
  est_ptrs_y,
  pipeline_x,
  pipeline_y,
  cor_cutoff,
  max_cor
)
}
\arguments{
\item{est_ptrs_x}{PTR estimates from sample subset x}

\item{est_ptrs_y}{PTR estimates from sample subset y}

\item{pipeline_x}{pipeline for sample subset x}

\item{pipeline_y}{pipeline for sample subset y}

\item{cor_cutoff}{the correlation cutoff}

\item{max_cor}{the max correlation}
}
\value{
a named list including the est_ptr dataframe and a max_cor value
\itemize{
\item sample: sample
\item est_ptr: PTR estimate
\item coefficient: coefficient of linear regression
\item pValue: p-value of linear regression
\item cor: correlation coefficient
\item correctY: corrected coverage
}
}
\description{
Compares sample subset x against sample subset y
}
