% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{compare_x_y}
\alias{compare_x_y}
\title{Compares contig subset x against contig subset y}
\usage{
compare_x_y(
  X,
  contig_subset_x,
  contig_subset_y,
  na_contig_ids,
  cor_cutoff,
  max_cor
)
}
\arguments{
\item{X}{input data frame}

\item{contig_subset_x}{the first set of contigs}

\item{contig_subset_y}{the second set of contigs}

\item{na_contig_ids}{the set of bad contigs discovered so far}

\item{cor_cutoff}{the correlation cutoff}

\item{max_cor}{the max correlation}
}
\value{
a dataframe
\itemize{
\item sample: sample
\item est_ptr: PTR estimate
\item coefficient: coefficient of linear regression
\item pValue: p-value of linear regression
\item cor: correlation coefficient
\item correctY: corrected coverage
}
}
\description{
Compares contig subset x against contig subset y
}
