% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{consist_transfer}
\alias{consist_transfer}
\title{A function for data frame integration}
\usage{
consist_transfer(x, y, i)
}
\arguments{
\item{x}{first data frame}

\item{y}{second data frame}

\item{i}{'sample' column}
}
\value{
a data frame with the other column as mean or max of that in the original two
}
\description{
A function for data frame integration
}
