% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{contig_pca}
\alias{contig_pca}
\title{A function to return the first dimension of PCA on an input matrix}
\usage{
contig_pca(X)
}
\arguments{
\item{X}{a matrix to undergo PCA}
}
\value{
first dimension of the PCA results
}
\description{
A function to return the first dimension of PCA on an input matrix
}
