% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ContigCluster2}
\alias{ContigCluster2}
\title{Contig Cluster 2}
\format{
\subsection{\code{ContigCluster2}}{

A data frame with 66,735 rows and 5 columns:
\describe{
\item{log_cov}{Log Coverage for Sliding Windows over Contigs}
\item{GC_content}{GC Content for Sliding Windows over Contigs}
\item{sample}{Sample Name}
\item{contig}{Contig Name}
\item{length}{Length of Contig}
}
}
}
\source{
\url{https://sourceforge.net/projects/demic/files/}
}
\usage{
ContigCluster2
}
\description{
Data associated with DEMIC paper (on SourceForge)
}
\keyword{datasets}
