% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPTR.R
\name{contigsPipeline}
\alias{contigsPipeline}
\title{Attempt the default iteration for contigs
Requires at least 20 contigs}
\usage{
contigsPipeline(X)
}
\arguments{
\item{X}{cov3 matrix as a dataframe}
}
\value{
estPTRs matrix on success, NULL otherwise
}
\description{
Attempt the default iteration for contigs
Requires at least 20 contigs
}
