% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_diff}
\alias{cor_diff}
\title{A function for orientation determination}
\usage{
cor_diff(Z)
}
\arguments{
\item{Z}{a vector of values}
}
\value{
a minor subset, where each value has the same orientation
}
\description{
A function for orientation determination
}
