% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGrowthRate1.0.2.R
\name{countNA}
\alias{countNA}
\title{A helper function for getting the sum and length of a vector}
\usage{
countNA(x)
}
\arguments{
\item{x}{the vector}
}
\value{
the vector containing the sum and length of the input
}
\description{
A helper function for getting the sum and length of a vector
}
