% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGrowthRate1.0.2.R
\name{demic}
\alias{demic}
\title{The main runner function for the DEMIC package}
\usage{
demic(
  bam_dir,
  fasta_dir,
  out_dir,
  window_size,
  window_step,
  mapq_cutoff,
  mapl_cutoff,
  max_mismatch_ratio,
  max_candidate_iter,
  max_threads
)
}
\arguments{
\item{bam_dir}{directory of input BAM files generated by bowtie2 (required)}

\item{fasta_dir}{directory of input FASTA files, each for a species (required)}

\item{out_dir}{directory of output files (default: ./)}

\item{window_size}{size (nt) of window for calculation of coverage (default: 5000)}

\item{window_step}{step (nt) of window for calculation of coverage (default: 100)}

\item{mapq_cutoff}{cutoff of mapping quality when calculating coverages (default: 5)}

\item{mapl_cutoff}{cutoff of mapping length when calculating coverages (default: 50)}

\item{max_mismatch_ratio}{maximum of mismatch ratio for each read as a hit (default: 0.03)}

\item{max_candidate_iter}{SOMETHING (default: 20)}
}
\description{
The main runner function for the DEMIC package
}
\examples{
demic("/input/bam_dir/", "/input/fa_dir/")
demic("/input/bam_dir/", "/input/fa_dir/", "/output/dir/")
}
