% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_transfer}
\alias{df_transfer}
\title{A function for data frame transfer}
\usage{
df_transfer(x, y)
}
\arguments{
\item{x}{first data frame with six columns}

\item{y}{second data frame with six columns}
}
\value{
a data frame with the same six columns but integrated info
}
\description{
A function for data frame transfer
}
