% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfTransfer}
\alias{dfTransfer}
\title{A function for data frame transfer}
\usage{
dfTransfer(X1, X2)
}
\arguments{
\item{X1}{first data frame with six columns}

\item{X2}{second data frame with six columns}
}
\value{
a data frame with the same six columns but integrated info
}
\description{
A function for data frame transfer
}
