% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ptr.R
\name{est_ptr}
\alias{est_ptr}
\title{Main function}
\usage{
est_ptr(X, max_candidate_iter = 10)
}
\arguments{
\item{X}{dataframe with coverage matrix
(column names: "logCov", "GC", "sample", "contig", "length")}

\item{max_candidate_iter}{max allowed iterations for estimation of PTR
(default: 10)}
}
\value{
dataframe with the estimated PTRs
(column names: "estPTR", "coefficient", "pValue", "cor", "correctY")
}
\description{
Main function
}
