% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{est_ptr_from_contigs}
\alias{est_ptr_from_contigs}
\title{Tries up to max_attempts times to compare each permutation of random
subsets of contigs from X, and returns the PTR estimate if a valid one comes
back from the comparisons}
\usage{
est_ptr_from_contigs(X, max_attempts = 3, cor_cutoff = 0.98, num_subsets = 5)
}
\arguments{
\item{X}{cov3 dataframe}

\item{max_attempts}{max number of attempts to find a valid contig}

\item{cor_cutoff}{correlation cutoff between PTR estimates between two subsets of contigs}

\item{num_subsets}{number of subsets to split contigs into}
}
\value{
est_ptrs dataframe on success, a message otherwise
}
\description{
Tries up to max_attempts times to compare each permutation of random
subsets of contigs from X, and returns the PTR estimate if a valid one comes
back from the comparisons
}
\examples{
est_ptrs_001 <- est_ptr(ContigCluster1)
est_ptrs_001

}
