% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{est_ptr_from_samples}
\alias{est_ptr_from_samples}
\title{Tries up to max_attempts times to compare each permutation of random
subsets of contigs from X, and returns the PTR estimate if a valid one comes
back from the comparisons}
\usage{
est_ptr_from_samples(X, max_candidate_iter = 10)
}
\arguments{
\item{X}{cov3 dataframe}

\item{max_candidate_iter}{max number of tries for samples pipeline iteration}
}
\value{
est_ptrs dataframe on success, null otherwise
\itemize{
\item est_ptr: estimated PTR values
\item coefficient: coefficient of linear regression
\item pValue: p-value of linear regression
\item cor: correlation coefficient
\item correctY: corrected coverage
}
}
\description{
Tries up to max_attempts times to compare each permutation of random
subsets of contigs from X, and returns the PTR estimate if a valid one comes
back from the comparisons
}
\examples{
est_ptrs_001 <- est_ptr_from_samples(ContigCluster1)
est_ptrs_001

}
