% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{est_ptr_on_all}
\alias{est_ptr_on_all}
\title{Estimates PTRs based on the whole input dataset}
\usage{
est_ptr_on_all(X)
}
\arguments{
\item{X}{cov3 dataframe}
}
\value{
est_ptrs dataframe on success, null otherwise
\itemize{
\item est_ptr: estimated PTR values
\item coefficient: coefficient of linear regression
\item pValue: p-value of linear regression
\item cor: correlation coefficient
\item correctY: corrected coverage
}
}
\description{
Estimates PTRs based on the whole input dataset
}
\examples{
est_ptrs_001 <- est_ptr_on_all(max_bin_003)
est_ptrs_001

}
