% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{est_ptrs_subset}
\alias{est_ptrs_subset}
\title{Get PTR estimates for output of the core pipeline on a subset of data}
\usage{
est_ptrs_subset(p)
}
\arguments{
\item{p}{is the pipeline named list}
}
\value{
a dataframe
\itemize{
\item sample: sample
\item est_ptr: PTR estimate
\item coefficient: coefficient of linear regression
\item pValue: p-value of linear regression
\item cor: correlation coefficient
\item correctY: corrected coverage
}
}
\description{
Get PTR estimates for output of the core pipeline on a subset of data
}
