% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGrowthRate1.0.2.R
\name{estGrowthRate}
\alias{estGrowthRate}
\title{Main function}
\usage{
estGrowthRate(input, output, max_candidate_iter, log_level)
}
\arguments{
\item{input}{path to the coverage matrix file in csv form (column names: "logCov", "GC", "sample", "contig", "length")}

\item{output}{path to the output directory (default: "getwd()/output/")}

\item{max_candidate_iter}{max allowed iterations for estimation of PTR (default: 10)}

\item{log_level}{logger threshold \link{TRACE, DEBUG, INFO, WARN, ERROR} (default: INFO)}
}
\value{
nothing, but outputs file with results
}
\description{
Main function
}
