% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPTR.R
\name{estPTR}
\alias{estPTR}
\title{Main function}
\usage{
estPTR(
  input,
  ...,
  output = file.path(getwd(), output),
  max_candidate_iter = 10,
  log_level = INFO
)
}
\arguments{
\item{input}{path to the coverage matrix file in csv form (column names: "logCov", "GC", "sample", "contig", "length")}

\item{output}{path to the output directory (default: "getwd()/output/")}

\item{max_candidate_iter}{max allowed iterations for estimation of PTR (default: 10)}

\item{log_level}{logger threshold \link{TRACE, DEBUG, INFO, WARN, ERROR} (default: INFO)}
}
\value{
nothing, but creates file with results
}
\description{
Main function
}
