% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_sample}
\alias{filter_sample}
\title{A function for sample filtration
Input requirements: 1. have values in more than half of the contigs 2. average log2(cov) > 0 in all these contigs}
\usage{
filter_sample(Z, avg_cutoff, cutoff_ratio)
}
\arguments{
\item{Z}{a matrix}

\item{avg_cutoff}{threshold of average}

\item{cutoff_ratio}{threshold of ratio}
}
\value{
the coefficient and p value of linear regression
}
\description{
A function for sample filtration
Input requirements: 1. have values in more than half of the contigs 2. average log2(cov) > 0 in all these contigs
}
