% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterSample}
\alias{filterSample}
\title{A function for sample filtration}
\usage{
filterSample(Z, cutoffAve, cutoffRatio)
}
\arguments{
\item{Z}{a matrix}

\item{cutoffAve}{threshold of average}

\item{cutoffRatio}{threshold of ratio}
}
\value{
the coefficient and p value of linear regression
Input requirements: 1. have values in more than half of the contigs 2. average log2(cov) > 0 in all these contigs
}
\description{
A function for sample filtration
}
