% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGrowthRate1.0.2.R
\name{itePipelines}
\alias{itePipelines}
\title{A function for iteration of pipeline until convergence}
\usage{
itePipelines(Z)
}
\arguments{
\item{Z}{a matrix of coverages}
}
\value{
final filtered samples, matrix of sample, contig and corrected coverages,
filtered contigs with PC1 values, PC1 range, preliminary filtered samples
}
\description{
A function for iteration of pipeline until convergence
}
