% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_core.R
\name{iterate_pipelines}
\alias{iterate_pipelines}
\title{A function for iteration of pipeline until convergence}
\usage{
iterate_pipelines(Z)
}
\arguments{
\item{Z}{a matrix of coverages}
}
\value{
final filtered samples, matrix of sample, contig and corrected coverages,
filtered contigs with PC1 values, PC1 range, preliminary filtered samples
}
\description{
A function for iteration of pipeline until convergence
}
