% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ks}
\alias{ks}
\title{A convenient function for KS test of uniform distribution}
\usage{
ks(x)
}
\arguments{
\item{x}{a vector without NA}
}
\value{
the p value of KS test
}
\description{
A convenient function for KS test of uniform distribution
}
