% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lm_column}
\alias{lm_column}
\title{A convenient function for ordinary linear regression on two vectors}
\usage{
lm_column(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\value{
the coefficient and p value of linear regression
}
\description{
A convenient function for ordinary linear regression on two vectors
}
