% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGrowthRate1.0.2.R
\name{lmColumn}
\alias{lmColumn}
\title{A convenient function for ordinary linear regression on two vectors}
\usage{
lmColumn(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\value{
the coefficient and p value of linear regression
}
\description{
A convenient function for ordinary linear regression on two vectors
}
