% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_core.R
\name{lme4_model}
\alias{lme4_model}
\title{Run mixed linear model with random effect using lme4}
\usage{
lme4_model(X)
}
\arguments{
\item{X}{input data frame}
}
\value{
a dataframe
}
\description{
Run mixed linear model with random effect using lme4
}
