% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{lme_model}
\alias{lme_model}
\title{Run mixed linear model with random effect}
\usage{
lme_model(X)
}
\arguments{
\item{X}{input data frame}
}
\value{
a dataframe
}
\description{
Run mixed linear model with random effect
}
