% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{max_bin_003}
\alias{max_bin_003}
\title{MaxBin2 Cluster 003}
\format{
\subsection{\code{max_bin_003}}{

A data frame with 124,578 rows and 5 columns:
\describe{
\item{log_cov}{Log Coverage for Sliding Windows over Contigs}
\item{GC_content}{GC Content for Sliding Windows over Contigs}
\item{sample}{Sample Name}
\item{contig}{Contig Name}
\item{length}{Length of Contig}
}
}
}
\source{
\url{https://sourceforge.net/projects/demic/files/}
}
\usage{
max_bin_003
}
\description{
Generated by PyCov3 on simulated test data
}
\keyword{datasets}
