% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGrowthRate.R, R/estGrowthRate1.0.2.R
\name{pipeline}
\alias{pipeline}
\title{A function representing the pipeline of four steps
including GC bias correction, sample filtration, PCA and contig filtration}
\usage{
pipeline(Y, i)

pipeline(Y, i)
}
\arguments{
\item{Y}{a matrix of coverages}

\item{i}{cutoff of filtering samples changes according to parameter i; i=1, cutoffRatio is 0.5; i=2, cutoffRatio is 1 as contig is clean}
}
\value{
is a named list containing \itemize{
\item \code{SamplesFilteredFinal} are the final filtered samples
\item \code{summaryMeanYSortFilteredSampleContig} is a matrix of samples
\item \code{contigPCAPC1Filtered} contig and corrected coverages, filtered contigs with PC1 values
\item \code{rangeVal} PC1 range
\item \code{Samples_filteredY} preliminary filtered samples
}

final filtered samples, matrix of sample, contig and corrected coverages,
filtered contigs with PC1 values, PC1 range, preliminary filtered samples
}
\description{
A function representing the pipeline of four steps
including GC bias correction, sample filtration, PCA and contig filtration

A function representing the pipeline of four steps
including GC bias correction, sample filtration, PCA and contig filtration
}
