% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{rand_ordered_contigs}
\alias{rand_ordered_contigs}
\title{Gives a randomly ordered set of contigs}
\usage{
rand_ordered_contigs(X)
}
\arguments{
\item{X}{input data frame that includes contigs}
}
\value{
contigs randomly ordered
}
\description{
Gives a randomly ordered set of contigs
}
