% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cov3.R
\name{read_cov3}
\alias{read_cov3}
\title{Read a PyCov3 coverage report}
\usage{
read_cov3(file, ...)
}
\arguments{
\item{file}{Path to a .cov3 file}

\item{...}{Additional arguments passed to readr::read_csv}
}
\value{
A tibble with columns \code{log_cov}, \code{GC_content}, \code{sample}, \code{contig},
and \code{length}.
}
\description{
Convenient helper around readr::read_csv that converts sample and contig
columns to factors.
}
\examples{
tmp <- tempfile(fileext = ".cov3")
write.csv(max_bin_001, tmp, row.names = FALSE, quote = FALSE)
read_cov3(tmp)

}
