% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reshapeFiltered}
\alias{reshapeFiltered}
\title{A function for reshape to facilitate PCA, removing all contigs with missing values for designated samples}
\usage{
reshapeFiltered(Samples_filtered, Z)
}
\arguments{
\item{Samples_filtered}{a vector of samples}

\item{Z}{a matrix of coverage}
}
\value{
a reshaped matrix of coverage
}
\description{
A function for reshape to facilitate PCA, removing all contigs with missing values for designated samples
}
