% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{samples_pipeline}
\alias{samples_pipeline}
\title{Attempt alternative iteration for samples
Requires at least 3 samples}
\usage{
samples_pipeline(X, max_candidate_iter)
}
\arguments{
\item{X}{cov3 matrix as a dataframe}

\item{max_candidate_iter}{max number of tries for samples pipeline iteration}
}
\value{
est_PTRs matrix on success, NULL otherwise
}
\description{
Attempt alternative iteration for samples
Requires at least 3 samples
}
