% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_by_ks_test}
\alias{select_by_ks_test}
\title{A function to remove outlier contigs using KS test}
\usage{
select_by_ks_test(sort_values)
}
\arguments{
\item{sort_values}{a vector of sorted values}
}
\value{
a vector with all values following a uniform distribution
}
\description{
A function to remove outlier contigs using KS test
}
