% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGrowthRate1.0.2.R
\name{selectAccordingToKSTest}
\alias{selectAccordingToKSTest}
\title{A function to remove outlier contigs using KS test}
\usage{
selectAccordingToKSTest(sortValues)
}
\arguments{
\item{sortValues}{a vector of sorted values}
}
\value{
a vector with all values following a uniform distribution
}
\description{
A function to remove outlier contigs using KS test
}
