% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_reasonable}
\alias{test_reasonable}
\title{A function to test whether the result is reasonable}
\usage{
test_reasonable(a, b)
}
\arguments{
\item{a}{first vector of values}

\item{b}{second vector of values}
}
\value{
the test result
}
\description{
A function to test whether the result is reasonable
}
