% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_PTR.R
\name{verify_input}
\alias{verify_input}
\title{Verify that the input dataframe/matrix is valid}
\usage{
verify_input(X)
}
\arguments{
\item{X}{dataframe/matrix with cov3 information}
}
\description{
Verify that the input dataframe/matrix is valid
}
