% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{contigs_pipeline}
\alias{contigs_pipeline}
\title{Attempt the default iteration for contigs
Requires at least 20 contigs}
\usage{
contigs_pipeline(X)
}
\arguments{
\item{X}{cov3 dataframe}
}
\value{
est_ptrs dataframe on success, a message otherwise
}
\description{
Attempt the default iteration for contigs
Requires at least 20 contigs
}
