% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_diff}
\alias{cor_diff}
\title{Determine the majority orientation of the input PTR estimates correlations}
\usage{
cor_diff(Z)
}
\arguments{
\item{Z}{a vector of values}
}
\value{
a minor subset, where each value has the same orientation
}
\description{
Determine the majority orientation of the input PTR estimates correlations
}
