% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_PTR.R
\name{est_ptr}
\alias{est_ptr}
\title{Main function}
\usage{
est_ptr(X, max_candidate_iter = 10)
}
\arguments{
\item{X}{dataframe with coverage matrix
(column names: "logCov", "GC", "sample", "contig", "length")}

\item{max_candidate_iter}{max allowed iterations for estimation of PTR
(default: 10)}
}
\value{
dataframe with the estimated PTRs
\itemize{
\item estPTR: estimated PTR values
\item coefficient: coefficient of linear regression
\item pValue: p-value of linear regression
\item cor: correlation coefficient
\item correctY: corrected coverage
}
}
\description{
Main function
}
\examples{
est_ptrs_001 <- est_ptr(max_bin_001)
est_ptrs_001

}
