% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_PTR.R
\name{est_ptr}
\alias{est_ptr}
\title{Estimate PTRs using all input data as well as using subsets of contigs and samples}
\usage{
est_ptr(X)
}
\arguments{
\item{X}{dataframe with coverage matrix
(column names: "log_cov", "GC_content", "sample", "contig", "length")}
}
\value{
named list with results from all three methods
all_ptr dataframe with the estimated PTRs on success, null otherwise
\itemize{
\item est_ptr: estimated PTR values
\item coefficient: coefficient of linear regression
\item pValue: p-value of linear regression
\item cor: correlation coefficient
\item correctY: corrected coverage
}
contigs_ptr dataframe with the estimated PTRs on success, null otherwise
\itemize{
\item est_ptr: estimated PTR values
\item coefficient: coefficient of linear regression
\item pValue: p-value of linear regression
\item cor: correlation coefficient
\item correctY: corrected coverage
}
samples_ptr dataframe with the estimated PTRs on success, null otherwise
\itemize{
\item est_ptr: estimated PTR values
\item coefficient: coefficient of linear regression
\item pValue: p-value of linear regression
\item cor: correlation coefficient
\item correctY: corrected coverage
}
}
\description{
Estimate PTRs using all input data as well as using subsets of contigs and samples
}
\examples{
est_ptrs_001 <- est_ptr(max_bin_003)
est_ptrs_001

}
