% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_PTR.R
\name{est_PTR}
\alias{est_PTR}
\title{Main function}
\usage{
est_PTR(X, max_candidate_iter = 10)
}
\arguments{
\item{X}{dataframe with coverage matrix (column names: "logCov", "GC", "sample", "contig", "length")}

\item{max_candidate_iter}{max allowed iterations for estimation of PTR (default: 10)}
}
\value{
dataframe with the estimated PTRs (column names: "estPTR", "coefficient", "pValue", "cor", "correctY")
}
\description{
Main function
}
