% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{est_ptr_from_samples}
\alias{est_ptr_from_samples}
\title{Attempt alternative iteration for samples
Requires at least 3 samples}
\usage{
est_ptr_from_samples(X, max_candidate_iter = 10)
}
\arguments{
\item{X}{cov3 dataframe}

\item{max_candidate_iter}{max number of tries for samples pipeline iteration}
}
\value{
est_ptrs dataframe
}
\description{
Attempt alternative iteration for samples
Requires at least 3 samples
}
\examples{
est_ptrs_001 <- est_ptr(ContigCluster1)
est_ptrs_001

}
