% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGrowthRate1.0.2.R
\name{estGrowthRate}
\alias{estGrowthRate}
\title{Main function}
\usage{
estGrowthRate(input, output, max_candidate_iter)
}
\arguments{
\item{input}{path to the coverage matrix file in csv form (column names: "logCov", "GC", "sample", "contig", "length")}

\item{output}{path to the output directory (default: "getwd()/output/")}

\item{max_candidtate_iter}{SOMETHING (default: 10)}
}
\value{
nothing, but outputs file with results
}
\description{
Main function
}
\examples{
estGrowthRate("/path/to/input/coverage.cov3", "path/to/output/", 10)
estGrowthRate("/path/to/input/coverage.cov3")
}
