% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_core.R
\name{iterate_pipelines}
\alias{iterate_pipelines}
\title{A function for iteration of pipeline until convergence}
\usage{
iterate_pipelines(Z)
}
\arguments{
\item{Z}{a matrix of coverages}
}
\value{
a named list
\itemize{
\item samples: vector of final filtered samples
\item correct_ys: matrix of sample, contig and corrected coverages
\item pc1: matrix of contig and PC1 values
\item pc1_range: vector of PC1 range
\item samples_y: samples filtered for reliable coverage
}
}
\description{
A function for iteration of pipeline until convergence
}
