% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_core.R
\name{pipeline}
\alias{pipeline}
\title{A function representing the pipeline of four steps
including GC bias correction, sample filtration, PCA and contig filtration}
\usage{
pipeline(Y, i)
}
\arguments{
\item{Y}{a matrix of coverages}

\item{i}{cutoff of filtering samples changes according to parameter i; i=1, cutoffRatio is 0.5; i=2, cutoffRatio is 1 as contig is clean}
}
\value{
a named list
\itemize{
\item samples: final list of filtered samples
\item correct_ys: dataframe with correct Y values per contig/sample
\item pc1: PC1 results of PCA per contig
\item pc1_range: range of PC1
\item samples_y: samples filtered for reliable coverage
}
}
\description{
A function representing the pipeline of four steps
including GC bias correction, sample filtration, PCA and contig filtration
}
