% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{samples_pipeline}
\alias{samples_pipeline}
\title{Attempt alternative iteration for samples
Requires at least 3 samples}
\usage{
samples_pipeline(X, max_candidate_iter = 10, contigs_pipeline_msg = "")
}
\arguments{
\item{X}{cov3 dataframe}

\item{max_candidate_iter}{max number of tries for samples pipeline iteration}

\item{contigs_pipeline_msg}{message from contigs_pipeline failure}
}
\value{
est_ptrs dataframe
}
\description{
Attempt alternative iteration for samples
Requires at least 3 samples
}
