% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_utils.R
\name{verify_input}
\alias{verify_input}
\title{Verify input data meets pipeline criteria}
\usage{
verify_input(X)
}
\arguments{
\item{X}{cov3 matrix as a dataframe}
}
\value{
TRUE if the input data meets the criteria, FALSE otherwise
}
\description{
Verify input data meets pipeline criteria
}
